requirements.txt
alembic==1.4.2
appdirs==1.4.3
attrs==19.3.0
Babel==2.8.0
black==19.10b0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dominate==2.5.1
elasticsearch==7.7.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.8.0
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
regex==2020.4.4
requests==2.23.0
rq==1.4.2
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.2.1
